minor: rename utils lib
