+++ src/component/Record.svelte
@@ -1,7 +1,7 @@
 <script>
 	import { recorder, } from '@lib/recorder.svelte.js';
 	import { session } from '@lib/config.svelte.js';
-	import { get_time_string, gen_href, } from '@lib/u.js';
+	import { get_time_string, gen_href, } from '@lib/u.svelte.js';
 	import { i18n } from '@lib/i18n.svelte.js';
 	import { pokemonStore, } from '@lib/pm.svelte.js';
