+++ src/component/Share.svelte
@@ -2,7 +2,7 @@
 	import { i18n } from '@lib/i18n.svelte.js';
 	import { session } from '@lib/config.svelte.js';
 	import { pokemonStore, } from '@lib/pm.svelte.js';
-	import { fetch_data, gen_href, } from '@lib/u.js';
+	import { fetch_data, gen_href, } from '@lib/u.svelte.js';
 
 	let url = $derived(gen_href(pokemonStore.status_string, session.name));
